requirements.txt
python-dotenv
paramiko
structlog